+++ Sources/EpoxyBars/BarModel/BarModel.swift
@@ -157,6 +157,10 @@ extension BarModel: StyleIDProviding {}
 
 extension BarModel: WillDisplayProviding {}
 
+// MARK: DidSelectProviding
+
+extension BarModel: DidSelectProviding {}
+
 // MARK: BarModeling
 
 extension BarModel: BarModeling {
@@ -166,6 +170,10 @@ extension BarModel: BarModeling {
 // MARK: InternalBarModeling
 
 extension BarModel: InternalBarModeling {
+  var isSelectable: Bool {
+    didSelect != nil
+  }
+
   func makeConfiguredView(traitCollection: UITraitCollection) -> UIView {
     let view = makeView()
     let context = CallbackContext(view: view, traitCollection: traitCollection, animated: false)
@@ -189,6 +197,11 @@ extension BarModel: InternalBarModeling {
   func didDisplay(_ view: UIView, traitCollection: UITraitCollection, animated: Bool) {
     didDisplay?(.init(view: castOrAssert(view), traitCollection: traitCollection, animated: animated))
   }
+
+  func didSelect(_ view: UIView, traitCollection: UITraitCollection, animated: Bool) {
+    didSelect?(.init(view: castOrAssert(view), traitCollection: traitCollection, animated: animated))
+  }
+
 }
 
 // MARK: InternalBarCoordinating