+++ Sources/EpoxyBars/BarModel/InternalBarModeling.swift
@@ -24,6 +24,12 @@ protocol InternalBarModeling: Diffable, EpoxyModeled {
 
   /// Should inform consumers that this bar model has been displayed.
   func didDisplay(_ view: UIView, traitCollection: UITraitCollection, animated: Bool)
+
+  /// Should inform consumers that this bar model has been selected.
+  func didSelect(_ view: UIView, traitCollection: UITraitCollection, animated: Bool)
+
+  /// Whether the bar model should be selectable.
+  var isSelectable: Bool { get }
 }
 
 // MARK: - InternalBarCoordinating