+++ Sources/EpoxyBars/TopBarInstaller/TopBarContainer.swift
@@ -15,10 +15,14 @@ public final class TopBarContainer: BarStackView, InternalBarContainer {
   // MARK: Lifecycle
 
   public init(didUpdateCoordinator: ((AnyBarCoordinating) -> Void)?) {
-    super.init(zOrder: .topToBottom, didUpdateCoordinator: didUpdateCoordinator)
+    super.init(style: .topToBottom, didUpdateCoordinator: didUpdateCoordinator)
 
     addSubviews()
     constrainSubviews()
+  }
+
+  required public init(style: Style) {
+    fatalError("init(style:) has not been implemented")
   }
 
   // MARK: Public